requirements.txt
livekit-agents>=0.8.0
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
supabase
